Refactoring (#17)

+++ scenes/app_settings/totp_app_settings.h
@@ -7,7 +7,7 @@
 #include "../../types/plugin_event.h"
 
 typedef struct {
-    uint8_t current_token_index;
+    uint16_t current_token_index;
 } AppSettingsSceneContext;
 
 void totp_scene_app_settings_init(const PluginState* plugin_state);
@@ -15,6 +15,8 @@ void totp_scene_app_settings_activate(
     PluginState* plugin_state,
     const AppSettingsSceneContext* context);
 void totp_scene_app_settings_render(Canvas* const canvas, PluginState* plugin_state);
-bool totp_scene_app_settings_handle_event(const PluginEvent* const event, PluginState* plugin_state);
+bool totp_scene_app_settings_handle_event(
+    const PluginEvent* const event,
+    PluginState* plugin_state);
 void totp_scene_app_settings_deactivate(PluginState* plugin_state);
 void totp_scene_app_settings_free(const PluginState* plugin_state);
